create validators module and test dir, rename utils.py to index.py
