+++ tests/validators/test_index.py
@@ -1,7 +1,6 @@
-
 import pytest
 
-from join_github_app.main.middleware.utils import is_valid_email_pattern
+from join_github_app.main.validators.index import is_valid_email_pattern
 
 class TestUtils:
